change: Replace IrisApi using methodhandles
Previously, we transitioned from using reflection to the `IrisApi` interface for accessing Iris functionality. However, since `IrisApi` is not available separately and can only be used after the release day as a major version, we have decided to switch back to using reflection.

Relying on the availability of `IrisApi` has introduced significant delays and dependencies on Iris. Until `IrisApi` becomes available separately, we will continue using reflection for accessing the required functionality.

This commit reverts the changes made to utilize `IrisApi` and restores the original reflection-based approach. Please note that this revert will not be backported to previous versions.